Fresnel propagator for both CUDA and CPU machines
first draft of the Fresnel propagator that auto detects if your machine
is GPU/CUDA enabled. Uses GPU if available and CPU otherwise. Needs
some speed improvements for the CPU version. In particular, if it is
run to a single propagation distance (i.e. Z is single valued and not a
vector).